requirements.txt
crewai
langchain-groq
crewai-tools
python-dotenv
fastapi
uvicorn
langchain_community